requirements.txt
gkeepapi
websockets
python-dotenv
